Twitter-API/Flask with Twitter Api/requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
future==0.18.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
oauthlib==3.1.0
python-twitter==3.5
requests==2.22.0
requests-oauthlib==1.3.0
urllib3==1.25.7
Werkzeug==0.16.0
